feat(desktop): replace Engine settings with a unified Hosts panel
Settings→Engine becomes Settings→Hosts: one panel with a This-device card
(engine status, base URL, Restart, this-device API key reveal/copy), an
Add-host row (URL + optional key + Test connection + single Add via
hosts.connect), a Connected list (disconnect + two-step Forget), a
Remembered list (Connect with the per-host key + Forget), and an On your
network list (instance-id-deduped discovery, one-click Add, THIS DEVICE
badge). No primary switch / 'Use this host' anywhere.

- settingsSchema: rename the first section Engine -> Hosts (kept default);
  models_dir/output_dir move to the hosts section.
- NavRail: drop the 'Switch to built-in engine' context entry; toast copy
  now points to Settings -> Hosts.
- CommandPalette: 'Set remote host…' -> 'Add host…' opening the Hosts panel.
- RunPodView: Use-in-Mold now hosts.connect(pod) + sticky generate target.
- EngineSection.vue/.test.ts removed; test ported to HostsSection.test.ts
  (reveal/copy hooks, no-Keychain, badges, two-step forget, per-host-secret
  connect) plus new Add-flow and instance-id discovery-dedupe tests.
